Air shuffles should not use fringe destination as scratch https://bugs.webkit.org/show_bug.cgi?id=283641
Reviewed by NOBODY (OOPS!).

When we have a shuffle in air, we have a series of shifts
and rotates. Rotates have extra bits not part of the loop,
called the fringe.

A series of shifts permits the destination of the first shift
to be used as scratch before it gets its final value. This is
not true for a fringe, even though a fringe is basically a shift.
This is only because the lifetime of the scratch register extends
into the rotate that follows it.

Basically, the mutable state in this code gets very confused
because of the constantly changing order of instructions emitted.

This bug was discovered with ARMv7 wasm tail calls, and the test case
will be included in that patch.

* Source/JavaScriptCore/b3/air/AirEmitShuffle.cpp:
(JSC::B3::Air::emitShuffle):