AX: In AccessibilityObject::visibleCharacterRange, we can sometimes loop infinitely over next-line-end positions https://bugs.webkit.org/show_bug.cgi?id=283640 rdar://140487777
Reviewed by Chris Fleizach.

Based on bug reports, it seems that we can loop infinitely in `AccessibilityObject::visibleCharacterRange()` while trying
to the `nextLineEndPosition`. It would be better if we understood why this happened, but can't reproduce it yet, so for
now just break if `nextLineEndPosition` returns the original position. The ASSERT should help us find a webpage where
this reproduces.

* Source/WebCore/accessibility/AccessibilityObject.cpp:
(WebCore::AccessibilityObject::visibleCharacterRangeInternal const):

Canonical link: https://commits.webkit.org/287036@main