[ Sequoia ] 3x TestWebKitAPI.ResourceLoad* (api-tests) are constant failures rdar://136525714 https://bugs.webkit.org/show_bug.cgi?id=280202
Reviewed by NOBODY (OOPS!).

We relax a bit the URL request class check to allow both mutable ans immutable NS url request classes.
ResourceRequest can either wrap a NSURLRequest or recreate a NSMutableURLRequest, depending on how the request is sent through IPC or manipulated.
If ResourceLoad delegates were API, it might be worth ensuring that the exposed request is NSURLRequest.
But this is not needed for this API test.

* Tools/TestWebKitAPI/Tests/WebKitCocoa/ResourceLoadDelegate.mm:
(LoadInfo)):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/ResourceLoadStatistics.mm:
(TEST(ResourceLoadStatistics, StorageAccessOnRedirectSitesWithQuirk)):
(TEST(ResourceLoadStatistics, StorageAccessGrantMultipleSubFrameDomains)):