B3LowerInt64: correctly align split Int64 values passed on the stack https://bugs.webkit.org/show_bug.cgi?id=283163
Reviewed by Justin Michaud.

When we split an Int64 value, we need to make sure that the first
component is at an 8-byte-aligned address, as defined by the ABI. We do
this by inserting a dummy 32-bit argument when necessary. Unfortunately,
to keep track of the stack alignment, we also need to track FP regs,
which results in some duplication with marshallCCallArgumentImpl.

* Source/JavaScriptCore/b3/B3LowerInt64.cpp:

Canonical link: https://commits.webkit.org/287058@main