Fix Mouseleave fires on elements no longer in the DOM https://bugs.webkit.org/show_bug.cgi?id=156971
Reviewed by NOBODY (OOPS!).

It fixes the bug where a deleted element fires  mouseleave events when we have mouse pointer inside it and move away from the element, even it is deleted.
by checking if the removed element node is still in the document tree. This way matching  the behaving from other browsers.

* LayoutTests/fast/events/mouse-events-on-node-deletion-expected.txt: Added.
* LayoutTests/fast/events/mouse-events-on-node-deletion.html: Added.
* LayoutTests/platform/ios/TestExpectations:
* Source/WebCore/page/EventHandler.cpp:
(WebCore::EventHandler::updateMouseEventTargetNode):
* LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_lostpointercapture_for_disconnected_node-expected.txt: