Remove unnecessary rounding leading to subpixel issues in elementFromPoint hit testing
https://bugs.webkit.org/show_bug.cgi?id=284698
rdar://141496802

Reviewed by NOBODY (OOPS!).

This patch removes rounding from hit testing code which leads to sub pixel
issues in elementFromPoint API.

Above fix leads to failure in case of `transformed` elements, so this merges
below Blink fix, which uses `transformed bounding box` for this purpose:

Merge: https://chromium-review.googlesource.com/c/chromium/src/+/1891110

Hit testing uses a 1x1 rect (i.e., "bounding box" )at the hit test
location to test intersection with boxes in the tree. The current
implementation always uses a 1x1 bounding box in the local coordinate
space of the box being tested, which is incorrect -- for example, if the
box is scaled by 100x, then using a 1x1 bounding box in the local space
is equivalent to using a 100x100 bounding box in the viewport, and as a
result we hit the transformed box even if the hit test location is still
100px away from it.

This patch fixes the issue by also transforming the bounding box, so
that it is always equivalent to a 1x1 rect in the viewport.

* Source/WebCore/rendering/HitTestLocation.cpp:
(WebCore::HitTestLocation::HitTestLocation):
(WebCore::HitTestLocation::HitTestLocation): Introduce another helper function
* Source/WebCore/rendering/HitTestLocation.h: Helper function definition
* Source/WebCore/rendering/HitTestingTransformState.cpp:
(HitTestingTransformState::boundsOfMappedQuad):
* Source/WebCore/rendering/HitTestingTransformState.h:
* Source/WebCore/rendering/RenderLayer.cpp:
(hitTestLayerByApplyingTransform):
* LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/elementFromPoint-subpixel-expected.txt: