Assert when layout in subgrid is updated as per added test case https://bugs.webkit.org/show_bug.cgi?id=284862
rdar://137177436

Reviewed by NOBODY (OOPS!).

The cause of the issue and the fix:
In some cases, when the difference between the old and the new value is layout change, the diff was flagged as zero, causing the subgridDidChange not to flag a change required condition. The issue was fixed by checking for layout update between the old and new styles and marking it as a valid difference.

* LayoutTests/fast/css-grid-layout/layout-update-assert-expected.txt: Added.
* LayoutTests/fast/css-grid-layout/layout-update-assert.html: Added.
* Source/WebCore/rendering/RenderGrid.cpp:
(WebCore::RenderGrid::subgridDidChange const):
* Source/WebCore/rendering/style/RenderStyle.cpp:
(WebCore::rareDataChangeRequiresLayout):