WebCodecs VideoFrame visibleWidth and visibleHeight should check for rotation rdar://143480580 https://bugs.webkit.org/show_bug.cgi?id=286421
Reviewed by NOBODY (OOPS!).

As shown in https://w3c.github.io/webcodecs/#dom-videoframe-videoframe-data-init, visibleWidth and visibleHeight are post frame rotation.
We thus swap visibleWidth and visibleHeight in case of 90 and 270 rotation.

This is important as rendering of video frames or texture generation in canvas will do the rotation.

* LayoutTests/http/wpt/mediastream/mediastreamtrackprocessor-videoframe-rotation-expected.txt: Added.
* LayoutTests/http/wpt/mediastream/mediastreamtrackprocessor-videoframe-rotation.html: Added.
* LayoutTests/platform/glib/TestExpectations:
* LayoutTests/webrtc/video-rotation-expected.txt:
* LayoutTests/webrtc/video-rotation.html:
* Source/WebCore/Modules/webcodecs/WebCodecsVideoFrame.cpp:
(WebCore::WebCodecsVideoFrame::create):
(WebCore::WebCodecsVideoFrame::initializeFrameFromOtherFrame):