Add out-of-flow objects under the inline in a continuation chain, when possible
https://bugs.webkit.org/show_bug.cgi?id=247681
rdar://102421379

Reviewed by NOBODY (OOPS!).

This patch aligns WebKit with Gecko / Firefox and Blink / Chromium.

Merge: https://chromium.googlesource.com/chromium/src/+/77ec19ddefb32688cd18277ed7412531b536401a

The same goes for floating objects. Only when a floating or out-of-flow
positioned object is to be added between two block-level children should we add
it to the anonymous block box holding the block-level children. If the new
child is to be added before a block-level child, and this beforeChild is the
first block-level child, we should rather make the new child the last child of
the preceding inline, rather than the first child of the anonymous block
containing block-level children. Also cleaned up and documented the code somewhat.

* Source/WebCore/rendering/updating/RenderTreeBuilderInline.cpp:
(WebCore::RenderTreeBuilder::Inline::insertChildToContinuation):
* LayoutTests/fast/inline/add-abspos-before-block-inside-inline.html:
* LayoutTests/fast/inline/add-abspos-before-block-inside-inline-expected.html: