Use drag recognizer in WebKit to send transform updates to Model View for Stage Mode behavior https://bugs.webkit.org/show_bug.cgi?id=287601 rdar://141251753
Reviewed by NOBODY (OOPS!).

In the third PR for stagemode, we implement the APIs in the WKStageModeInteractionDriver to
map XY translations from the UIPanGestureRecognizer to the rotation of the model, to create the orbit
interaction. This also adds additional API calls to and from the driver back to the ModelProcessModelPlayerProxy
such that the entityTransform in the HTMLModelElement is updated while the transform is occurring.

* Source/WebCore/page/ios/EventHandlerIOS.mm:
(WebCore::EventHandler::stageModeSessionDidUpdate):
* Source/WebKit/ModelProcess/cocoa/ModelProcessModelPlayerProxy.mm:
(WebKit::ModelProcessModelPlayerProxy::didFinishLoading):
* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView didReceiveInteractiveModelElement:]):
* Source/WebKit/WebKitSwift/StageMode/WKStageMode.swift:
(WKStageModeInteractionDriver.initialTargetPose):
(WKStageModeInteractionDriver.interactionDidBegin(_:)):
(WKStageModeInteractionDriver.interactionDidUpdate(_:)):
(WKStageModeInteractionDriver.operationDidChange(_:)):
- Adopt and update all the stage mode interaction APIs to update the model rotation based on the transform
updates from the pan gesture

* Source/WebKit/WebKitSwift/StageMode/WKStageModeBridging.h:
* Source/WebKit/WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::requestInteractiveModelElementAtPoint):
* Source/WebKit/WebProcess/WebPage/WebPage.h: