Hit-testing an element does not take the mask-image's bounds into account https://bugs.webkit.org/show_bug.cgi?id=287510
Reviewed by NOBODY (OOPS!).

Add a method to RenderBox, `hitTestMask`, allowing it's mask's bounds
to be hit tested, and take that into account when hit-testing a
RenderBox, if applicable. This avoids cursor events being dispatched
to masked-out portions of children of an element with a `mask-image`,
for instance. This behaviour appears not to be specified by
CSS-MASKING-1, but is consistent with what both Blink and Gecko
implement, and probably makes more sense from an user's perspective
either way.

* LayoutTests/css3/masking/mask-image-hit-test-children-outside-parent-bounds-expected.txt: Added.
* LayoutTests/css3/masking/mask-image-hit-test-children-outside-parent-bounds.html: Added.
* Source/WebCore/rendering/RenderBox.cpp:
(WebCore::RenderBox::hitTestMask const):

Add a method to hit test the RenderBox's mask's bounds, similar to
the border radius test immediately above.

(WebCore::RenderBox::maskClipRect const):
(WebCore::RenderBox::maskClipRect): Deleted.

Made `this` const, so we can make `hitTestMask`'s `this` const as well

(WebCore::RenderBox::nodeAtPoint):

Take the mask's bounds into account when deciding whether a given hit
has hit the current RenderBox.

* Source/WebCore/rendering/RenderBox.h:
* Source/WebCore/rendering/RenderLayer.cpp:
(WebCore::RenderLayer::hitTestLayer):

Likewise, for children of a given layer.