get_title() in LayoutTests/resources/testharness.js does not account for folders containing a full stop (.) in the name. https://bugs.webkit.org/show_bug.cgi?id=287285
Reviewed by NOBODY (OOPS!).

The implementation assumed that the pathname only contained one full stop character right before the extension.
If the pathname included a full stop in a folder, it would select a part of the folder structure as the test name.
As an example, the pathname `/a.folder/testname.html` would result in `get_title()` returning `.folder/` instead of `testname`..

The issue should be solved by splitting the operation into two steps, first trim away the path, and then the extension.
This change is copied from wpt upstream.

* LayoutTests/resources/testharness.js: