UnifiedPDF: GPU process exceeds memory limits when zooming https://bugs.webkit.org/show_bug.cgi?id=287478 rdar://144608457
Reviewed by NOBODY (OOPS!).

Mark the PDF layers as needing native context, so that they get
in process layer backing store.

Fixes cases where GPU process would exceed its memory limits because
CA would create temporary images for draws, and the temporary images
would not be attributed to the WebContent process. The temporary
images are likely needed due to layer tile and PDF image tile colorspace
mismatch.

The PDF images are rendered in WCP, and in theory there
is no benefit in transferring the data to GPUP for blitting
to a layer tile. GPU process transfer neccessitates a memcpy of the data.
The blit to a layer tile should be a memcpy. Thus keeping the data in
WCP should be optimal. However, due to inefficiency of the
PDF paint code, WCP will do a lot more work than a single memcpy. It
appears that the general case perf is still comparable to the
transfer to GPUP.

* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/PDFDiscretePresentationController.h:
* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/PDFDiscretePresentationController.mm:
(WebKit::PDFDiscretePresentationController::buildRows):
(WebKit::PDFDiscretePresentationController::layerNeedsPlatformContext const):
* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/PDFPresentationController.mm:
(WebKit::PDFPresentationController::makePageContainerLayer):
* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/PDFScrollingPresentationController.h:
* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/PDFScrollingPresentationController.mm:
(WebKit::PDFScrollingPresentationController::setupLayers):
(WebKit::PDFScrollingPresentationController::layerNeedsPlatformContext const):