Invalid IPC crash when a client provides an NSError with differing failing URL keys https://bugs.webkit.org/show_bug.cgi?id=288737 rdar://145681217
Reviewed by Ryosuke Niwa.

Clients can provide an NSError in API, which may be sent to the WCP/NP and may fail the IPC validation
added in rdar://140567340. For now, we should just implement a more targeted fix, protecting only the IPC
endpoint that we know is at risk when the keys do not match.

* Source/WebCore/platform/network/cf/ResourceError.h:
* Source/WebCore/platform/network/mac/ResourceErrorMac.mm:
(WebCore::ResourceError::hasMatchingFailingURLKeys const):
* Source/WebKit/Shared/Cocoa/CoreIPCError.mm:
(WebKit::CoreIPCError::hasValidUserInfo):
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::didFailProvisionalLoadForFrameShared):

Originally-landed-as: 289651.196@safari-7621-branch (113710100ae4). rdar://148056574