element.append() should only take the last occurrence of an element to insert it as a child https://bugs.webkit.org/show_bug.cgi?id=288955
Reviewed by NOBODY (OOPS!).

element.append() should only take the last occurrence of an element to insert it as a child.
In the current implementation, when append had duplicate values, it was adding the first occurrence of the element as a child.
This was causing a WPT test case to fail.

* LayoutTests/imported/w3c/web-platform-tests/dom/nodes/ParentNode-append-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/dom/nodes/ParentNode-append.html:
* Source/WebCore/dom/ContainerNode.cpp:
(WebCore::ContainerNode::append):