[Grid] Introduce grid area logical sizes cache. https://bugs.webkit.org/show_bug.cgi?id=289004 rdar://problem/146051830
Reviewed by NOBODY (OOPS!).

This patch introduces a cache that stores the logical size of each grid
item's grid area. The grid area of an item is computed during the track
sizing algorithm and is used during layoutGridItems to set the item's
grid area logical width/height before calling layout on it as this
represents the item's containing block. We can use this as an
opportunity to cache these computed values to potentially use in a
subsequent layout in certain conditions which will be determined in a
future patch.

With the introduction of this cache we will want to invalidate it
basically whenever the grid experiences any sort of style mutation that
requires layout or whenever the content of the grid changes. The former
is done in a straightforward manner via RenderGrid::styleDidChange. The
latter is done inside of dirtyGrid, which is called at different points
in time when we need to perform item placement again including when:

1. A renderer is attached or detached from the RenderGrid
2.) A grid item's style changes such that it requires it to be placed
again

We will also aggressively invalidate this cache at the beginning of
RenderGrid::layoutGrid since we have no use for it yet. With respect to
the scenarios described above, this would represent another case where
a grid item's content/styling would have been mutated and could possibly
influence the sizes of the tracks and the sizes of the grid areas as a
result. I plan to add a fast path in which we would not invalidate this
cache during layout and instead use the sizes in scenarios where the
grid content does not influence the sizes of the tracks.

* Source/WebCore/rendering/RenderGrid.cpp:
(WebCore::RenderGrid::styleDidChange):
(WebCore::RenderGrid::layoutGrid):
(WebCore::RenderGrid::layoutMasonry):
(WebCore::RenderGrid::dirtyGrid):
(WebCore::RenderGrid::layoutGridItems):
* Source/WebCore/rendering/RenderGrid.h: