non standard new Date('2024-12-3') yields to Invalid Date https://bugs.webkit.org/show_bug.cgi?id=283825 rdar://141044926
Reviewed by NOBODY (OOPS!).

This relaxes parsing rules for months and days to allow
dates such as 2025-3-16 for 2025-3-16 (16 March 2025)
or dates such as 2025-10-3 for 2025-10-03 (3 October 2025).

The constraint is a common request from web developers to
be more compatible with Firefox and Chrome behaviors.

So instead of Safari returning:

  new Date('2024-12-3') yields to: Invalid Date
  Date.parse('2024-12-3') yields to: NaN

It will now return:
  new Date('2024-12-3') yields to: Date Tue Dec 03 2024 00:00:00 GMT+0100 (Central European Standard Time)
  Date.parse('2024-12-3') yields to: 1733180400000

This is not conforming to the specification, but the specification
doesn't disallow it either.
https://tc39.es/ecma262/multipage/numbers-and-dates.html#sec-date-time-string-format

* Source/WTF/wtf/DateMath.cpp:
(WTF::parseES5DatePortion):