Stop Yarr from popping from an empty list https://bugs.webkit.org/show_bug.cgi?id=290010 rdar://146934667
Reviewed by NOBODY (OOPS!).

In 292003@main, we landed an optimization for Yarr for processing strings of pattern
characters broken up by character classes. However, there is a loop where we fail to
check the size of a list before calling removeLast on it, meaning that we may remove
a nonexistent element. This patch hardens that logic.

* Source/JavaScriptCore/yarr/YarrJIT.cpp: