GPUP: Drawing 2D context to a layer is slow due to CGImage creation moment https://bugs.webkit.org/show_bug.cgi?id=290168 rdar://147572105
Reviewed by NOBODY (OOPS!).

The CGImage for a particular 2D context canvas rendering would be
created during the process in which the canvas backing buffer is drawn
to the layer. It was done at last possible moment, e.g. at the
DrawImageBuffer call site.

For accelerated CG, creating an image will be flushing operation. This
may cause waiting on the already drawn layer content to be scheduled.

Instead, create images for all the updated 2d context canvas elements
during the "prepareForDisplay" phase. This way the flushes happen
with only the canvas draw calls in the queues, minimizing the
wait time.

* Source/WebCore/html/canvas/CanvasRenderingContext2DBase.cpp:
(WebCore::CanvasRenderingContext2DBase::needsPreparationForDisplay const):
* Source/WebCore/platform/graphics/ImageBuffer.cpp:
(WebCore::ImageBuffer::prepareForDisplay):
Avoid flushing here. Instead, prepareForDisplay should be an isolated
command: it should do whatever it needs, for the particular backend.
Before, it was only implemented for Skia, and that did not need
the flush.
* Source/WebCore/platform/graphics/ImageBuffer.h:
* Source/WebCore/platform/graphics/cg/ImageBufferIOSurfaceBackend.cpp:
(WebCore::ImageBufferIOSurfaceBackend::prepareForDisplay):
* Source/WebCore/platform/graphics/cg/ImageBufferIOSurfaceBackend.h:
* Source/WebKit/GPUProcess/graphics/RemoteImageBuffer.cpp:
(WebKit::RemoteImageBuffer::prepareForDisplay):
* Source/WebKit/GPUProcess/graphics/RemoteImageBuffer.h:
* Source/WebKit/GPUProcess/graphics/RemoteImageBuffer.messages.in:
* Source/WebKit/WebProcess/GPU/graphics/RemoteImageBufferProxy.cpp:
(WebKit::RemoteImageBufferProxy::prepareForDisplay):
* Source/WebKit/WebProcess/GPU/graphics/RemoteImageBufferProxy.h: