[ARMv7] Allow run-javascriptcore-tests to run arm tests on arm64 linux https://bugs.webkit.org/show_bug.cgi?id=290635
Reviewed by NOBODY (OOPS!).

On linux arm64 systems like the Thelio Astra, we can also run userspace armv7 binaries
with an aarch64 kernel. This patch makes the test runner accept this configuration,
without affecting x86 rosetta runs on macos. It is possible that we might run arm64
tests when we intend to run arm tests without an error now; the container should be carefully set up
to avoid this instead.

* Tools/Scripts/run-javascriptcore-tests:
(runJSCStressTests):
* Tools/Scripts/webkitdirs.pm:
(determineNativeArchitecture):
(determineArchitecture):