Yarr generatePatternCharacterOnce should limit how far away it coalesces characters https://bugs.webkit.org/show_bug.cgi?id=290789 rdar://148244655
Reviewed by NOBODY (OOPS!).

Currently, our optimization is able to read pattern characters any number of characters
away, as long as everything in between is empty. This could lead to unnecessary loading
of those intermediate characters. We constrain the JIT to search up to 16 characters in
each direction (two 8B loads), allowing us to group close pattern characters, while not
getting too aggressive with this optimization.

* Source/JavaScriptCore/yarr/YarrJIT.cpp: