[web-animations] `finish` animation event listener is not fired if registered in the `finished` promise https://bugs.webkit.org/show_bug.cgi?id=290836
Reviewed by NOBODY (OOPS!).

Dispatch of animation events is asynchronous. The relevant steps of the "update animations and send
events" procedure [0] for events is:

1. advance the current time of an animation and enqueue animation events (if any),
2. perform a microtask checkpoint,
3. dispatch any queued animation events.

In the case of an animation that naturally advances to its finished state, a `finish` event would be
enqueued in step 1, then the `finished` promise would yield JS code execution in step 2, and finally
the `finish` event enqueued in step 1 would be dispatched in step 3.

As part of https://commits.webkit.org/262123@main we made it so that we only enqueue `finish` events
if a `finish` event listener is currently registered. However, in the case where a `finish` event
listener is added after that event should have been enqueued, for instance during step 2 above,
then the `finish` event would not be dispatched.

We now remove this check and add an additional subtest to the WPT test devoted to testing the finishing
behavior of Web Animations in `web-animations/timing-model/animations/finishing-an-animation.html`.

Addressing this also brought to a light an issue in a Scroll-driven Animations WPT where we expected
that a `finished` promise resolving meant any pending `finish` event would have been dispatched. We
fix this test to wait another frame before making this assumption. That test passed in both Safari
and Chrome due to both browsers suffering from the issue being fixed here. A Chromium bug was filed [1].

[0] https://drafts.csswg.org/web-animations-1/#update-animations-and-send-events
[1] https://issues.chromium.org/issues/407497938

* LayoutTests/imported/w3c/web-platform-tests/scroll-animations/scroll-timelines/scroll-animation.html:
* LayoutTests/imported/w3c/web-platform-tests/web-animations/timing-model/animations/finishing-an-animation-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/web-animations/timing-model/animations/finishing-an-animation.html:
* Source/WebCore/animation/WebAnimation.cpp:
(WebCore::WebAnimation::finishNotificationSteps):